requirements.txt
pyside6
mysql-connector-python
python-dotenv